updated contacts controller

+++ controllers/contacts.js
@@ -1,18 +1,11 @@
 const asyncHandler = require("express-async-handler");
-const Contact = require("../models/contact");
+const Contact = require("../models/contacts");
 
-const getContact = asyncHandler(async (req, res) => {
-  const contact = await Contact.findById({ _id: req.params.id });
-  if (!contact) {
-    res.status(404);
-    throw new Error("Contact not found");
-  }
-  res.status(200).json(contact);
-});
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
+
 const createContact = asyncHandler(async (req, res) => {
   console.log("The request body is :", req.body);
   const { name, email, phone } = req.body;
@@ -24,15 +17,31 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
 
   res.status(201).json(contact);
 });
+
+const getContact = asyncHandler(async (req, res) => {
+  const contact = await Contact.findById(req.params.id);
+  if (!contact) {
+    res.status(404);
+    throw new Error("Contact not found");
+  }
+  res.status(200).json(contact);
+});
+
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
   }
 
   const updatedContact = await Contact.findByIdAndUpdate(
@@ -43,19 +52,25 @@ const updateContact = asyncHandler(async (req, res) => {
 
   res.status(200).json(updatedContact);
 });
+
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
   }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
+  }
   await Contact.deleteOne({ _id: req.params.id });
   res.status(200).json(contact);
 });
+
 module.exports = {
-  getContact,
   getContacts,
   createContact,
+  getContact,
   updateContact,
   deleteContact,
 };
